Story-2 Story-4 Add Dependency Injection using Dagger. Set up the View Models and Networking Layers on the DI Graph as well as the Application Object
Set Up Model Objects for Weather API call using RxJava

Add a Button to the Input Location Layout
Connect Value of the Button through the ViewModel Bindings to make an API call to look up the
nearest weather station to the requested text based location
An array of results is returned or an error message is displayed.
The top weather station is always accepted and the viewmodel observables are updated with the relevant station id
Then the app is triggered to navigate using the updated navigation graph to open the ForcastFragment

Upon binding the newly created Forcast Fragment and the ForcastViewModel, another API call is  fetched using the weather station result.

This fetches a list of WeatherDay objects containing a 5 day forcast and displays it in a recyclerview using data view binding in the adapter

Story 4. Integrated into the View, is a pull to refresh mechanism which clears the list of weather, updates the list and updates a timestamp when it was fetched.

The top panel also displays the Weather Station Name and Coordinates based on the text search.

In summary, the Go button now fetches the data from the server